+++ gaios/Shared/Controllers/DialogPassphrase/DialogPassphraseViewController.swift
@@ -70,7 +70,7 @@ class DialogPassphraseViewController: KeyboardViewController {
         let tapToClose = UITapGestureRecognizer(target: self, action: #selector(didTap))
             tappableBg.addGestureRecognizer(tapToClose)
 
-//        AnalyticsManager.shared.recordView(.requestAmount, sgmt: AnalyticsManager.shared.subAccSeg(AccountsRepository.shared.current, walletType: wallet?.type))
+//        AnalyticsManager.shared.recordView(.requestAmount, sgmt: AnalyticsManager.shared.subAccSeg(AccountsRepository.shared.current, walletItem: wallet?.type))
     }
 
     deinit {