+++ gaios/Utility/AssetExpandableSelect/AssetExpandableSelectViewController.swift
@@ -221,7 +221,7 @@ extension AssetExpandableSelectViewController: UITableViewDelegate, UITableViewD
             }
         }
         let account = viewModel.accountSelectSubCellModels[indexPath.row].account
-        AnalyticsManager.shared.selectAccount(account: AccountsRepository.shared.current, walletType: account.type)
+        AnalyticsManager.shared.selectAccount(account: AccountsRepository.shared.current, walletItem: account)
         AnalyticsManager.shared.selectAsset(account: AccountsRepository.shared.current)
         delegate?.didSelectReceiver(assetId: assetId, account: account)
         navigationController?.popViewController(animated: true)
@@ -273,7 +273,7 @@ extension AssetExpandableSelectViewController: UITextFieldDelegate {
 extension AssetExpandableSelectViewController: SecuritySelectViewControllerDelegate {
     func didCreatedWallet(_ wallet: WalletItem) {
 
-        AnalyticsManager.shared.createAccount(account: AccountsRepository.shared.current, walletType: wallet.type)
+        AnalyticsManager.shared.createAccount(account: AccountsRepository.shared.current, walletItem: wallet)
         delegate?.didSelectReceiver(assetId: getAssetId(), account: wallet)
         navigationController?.popViewController(animated: true)
     }