+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 - Improve QR code readability 
+- Update performance metrics
 
 ## [4.0.9] - 2023-06-29